amplify/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amplify"
version = "1.0.0"
description = "AMPLIFY: Actionless Motion Priors for Robot Learning from Videos"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jeremy A. Collins", email = "jer@gatech.edu"},
    {name = "Loránd Cheng", email = "lorand@gatech.edu"},
    {name = "Kunal Aneja", email = "kaneja6@gatech.edu"},
    {name = "Albert Wilcox", email = "albertwilcox@gatech.edu"},
    {name = "Benjamin Joffe", email = "benjamin.joffe@gtri.gatech.edu"},
    {name = "Animesh Garg", email = "animesh.garg@gatech.edu"}
]
keywords = [
    "robotics",
    "video modeling",
    "world modeling",
    "keypoint trajectory prediction",
    "imitation learning",
    "computer vision"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "bddl==1.0.1",
    "cloudpickle==2.1.0",
    "colorednoise==2.2.0",
    "easydict==1.9",
    "einops==0.8.0",
    "gym==0.25.2",
    "h5py==3.11.0",
    "huggingface-hub==0.24.6",
    "hydra-core==1.2.0",
    "imageio==2.35.1",
    "ipython==8.26.0",
    "matplotlib==3.5.3",
    "moviepy==1.0.3",
    "mujoco==3.2.2",
    "natsort==8.4.0",
    "numpy<2.0.0,>=1.20.0",
    "omegaconf==2.3.0",
    "opencv-python==4.6.0.66",
    "pillow==10.4.0",
    "positional-encodings==6.0.3",
    "pyinstrument==4.7.2",
    "requests==2.32.3",
    "robomimic==0.2.0",
    "robosuite==1.4.1",
    "safetensors==0.4.4",
    "scipy==1.13.1",
    "sentencepiece==0.2.0",
    "termcolor==2.4.0",
    "thop==0.1.1.post2209072238",
    "timm==1.0.8",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.66.5",
    "transformers==4.21.1",
    "vector-quantize-pytorch==1.20.11",
    "wandb==0.13.1",
]

[project.urls]
Homepage = "https://amplify-robotics.github.io/"
Repository = "https://github.com/pairlab/AMPLIFY"
Paper = "https://arxiv.org/pdf/2506.14198"
Documentation = "https://github.com/pairlab/AMPLIFY#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "isort>=5.10",
    "flake8>=5.0",
]

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["amplify*", "preprocessing*"]
exclude = [
    "LIBERO*",
    "co-tracker*",
    "preprocessed_data*",
    "outputs*",
    "wandb*",
    "logs*",
    "checkpoints*",
    "tests*",
]
